third_party/WebKit/LayoutTests/http/tests/security/xssAuditor/xss-filter-bypass-long-string-expected.txt

CONSOLE ERROR: line 79: The XSS Auditor refused to execute a script in 'http://localhost:8000/security/xssAuditor/resources/xss-filter-bypass-long-string-reply.php' because its source code was found within the request. The server sent an 'X-XSS-Protection' header requesting this behavior.
